sidebar major refactoring and cleanup
