+++ server/src/main/kotlin/net/horizonsend/ion/server/features/sidebar/component/ContactsSidebarComponent.kt
@@ -1,14 +1,13 @@
 package net.horizonsend.ion.server.features.sidebar.component
 
-import net.horizonsend.ion.server.features.sidebar.ContactsData
+import net.horizonsend.ion.server.features.sidebar.tasks.Contacts
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.Component.text
-//import net.kyori.adventure.text.format.NamedTextColor.DARK_GRAY
 import net.megavex.scoreboardlibrary.api.sidebar.component.LineDrawable
 import net.megavex.scoreboardlibrary.api.sidebar.component.SidebarComponent
 import java.util.function.Supplier
 
-class ContactsSidebarComponent(private val valueSupplier: Supplier<ContactsData>) : SidebarComponent {
+class ContactsSidebarComponent(private val valueSupplier: Supplier<Contacts.ContactsData>) : SidebarComponent {
 
     override fun draw(drawable: LineDrawable) {
         val value = valueSupplier.get()
